# sql/sqlstats: avoid heap allocations when using stats iterators