--- a/pkg/sql/sqlstats/persistedsqlstats/mem_iterator.go
+++ b/pkg/sql/sqlstats/persistedsqlstats/mem_iterator.go
@@ -25,7 +25,7 @@ import (
 // and aggregation_interval fields on the returning
 // appstatspb.CollectedStatementStatistics.
 type memStmtStatsIterator struct {
-	*sslocal.StmtStatsIterator
+	sslocal.StmtStatsIterator
 	aggregatedTs time.Time
 	aggInterval  time.Duration
 }
@@ -35,8 +35,8 @@ func newMemStmtStatsIterator(
 	options *sqlstats.IteratorOptions,
 	aggregatedTS time.Time,
 	aggInterval time.Duration,
-) *memStmtStatsIterator {
-	return &memStmtStatsIterator{
+) memStmtStatsIterator {
+	return memStmtStatsIterator{
 		StmtStatsIterator: stats.StmtStatsIterator(options),
 		aggregatedTs:      aggregatedTS,
 		aggInterval:       aggInterval,
@@ -59,7 +59,7 @@ func (m *memStmtStatsIterator) Cur() *appstatspb.CollectedStatementStatistics {
 // aggregatoin_interval fields fields on the returning
 // appstatspb.CollectedTransactionStatistics.
 type memTxnStatsIterator struct {
-	*sslocal.TxnStatsIterator
+	sslocal.TxnStatsIterator
 	aggregatedTs time.Time
 	aggInterval  time.Duration
 }
@@ -69,8 +69,8 @@ func newMemTxnStatsIterator(
 	options *sqlstats.IteratorOptions,
 	aggregatedTS time.Time,
 	aggInterval time.Duration,
-) *memTxnStatsIterator {
-	return &memTxnStatsIterator{
+) memTxnStatsIterator {
+	return memTxnStatsIterator{
 		TxnStatsIterator: stats.TxnStatsIterator(options),
 		aggregatedTs:     aggregatedTS,
 		aggInterval:      aggInterval,