--- a/pkg/sql/sqlstats/persistedsqlstats/stmt_reader.go
+++ b/pkg/sql/sqlstats/persistedsqlstats/stmt_reader.go
@@ -28,7 +28,7 @@ import (
 
 // IterateStatementStats implements sqlstats.Provider interface.
 func (s *PersistedSQLStats) IterateStatementStats(
-	ctx context.Context, options *sqlstats.IteratorOptions, visitor sqlstats.StatementVisitor,
+	ctx context.Context, options sqlstats.IteratorOptions, visitor sqlstats.StatementVisitor,
 ) (err error) {
 	// We override the sorting options since otherwise we would need to implement
 	// sorted and unsorted merge separately. We can revisit this decision if
@@ -79,7 +79,7 @@ func (s *PersistedSQLStats) IterateStatementStats(
 }
 
 func (s *PersistedSQLStats) persistedStmtStatsIter(
-	ctx context.Context, options *sqlstats.IteratorOptions,
+	ctx context.Context, options sqlstats.IteratorOptions,
 ) (iter isql.Rows, expectedColCnt int, err error) {
 	query, expectedColCnt := s.getFetchQueryForStmtStatsTable(ctx, options)
 
@@ -99,7 +99,7 @@ func (s *PersistedSQLStats) persistedStmtStatsIter(
 }
 
 func (s *PersistedSQLStats) getFetchQueryForStmtStatsTable(
-	ctx context.Context, options *sqlstats.IteratorOptions,
+	ctx context.Context, options sqlstats.IteratorOptions,
 ) (query string, colCnt int) {
 	selectedColumns := []string{
 		"aggregated_ts",