--- a/pkg/sql/sqlstats/sslocal/sslocal_iterator.go
+++ b/pkg/sql/sqlstats/sslocal/sslocal_iterator.go
@@ -27,14 +27,14 @@ type baseIterator struct {
 // statement statistics stored in SQLStats.
 type StmtStatsIterator struct {
 	baseIterator
-	curIter *ssmemstorage.StmtStatsIterator
+	curIter ssmemstorage.StmtStatsIterator
 }
 
 // NewStmtStatsIterator returns a new instance of the StmtStatsIterator.
-func NewStmtStatsIterator(s *SQLStats, options *sqlstats.IteratorOptions) *StmtStatsIterator {
+func NewStmtStatsIterator(s *SQLStats, options *sqlstats.IteratorOptions) StmtStatsIterator {
 	appNames := s.getAppNames(options.SortedAppNames)
 
-	return &StmtStatsIterator{
+	return StmtStatsIterator{
 		baseIterator: baseIterator{
 			sqlStats: s,
 			idx:      -1,
@@ -53,7 +53,7 @@ func NewStmtStatsIterator(s *SQLStats, options *sqlstats.IteratorOptions) *StmtS
 func (s *StmtStatsIterator) Next() bool {
 	// If we haven't called Next() for the first time or our current child
 	// iterator has finished iterator, then we increment s.idx.
-	if s.curIter == nil || !s.curIter.Next() {
+	if !s.curIter.Initialized() || !s.curIter.Next() {
 		s.idx++
 		if s.idx >= len(s.appNames) {
 			return false
@@ -78,14 +78,14 @@ func (s *StmtStatsIterator) Cur() *appstatspb.CollectedStatementStatistics {
 // statement statistics stored in SQLStats.
 type TxnStatsIterator struct {
 	baseIterator
-	curIter *ssmemstorage.TxnStatsIterator
+	curIter ssmemstorage.TxnStatsIterator
 }
 
 // NewTxnStatsIterator returns a new instance of the TxnStatsIterator.
-func NewTxnStatsIterator(s *SQLStats, options *sqlstats.IteratorOptions) *TxnStatsIterator {
+func NewTxnStatsIterator(s *SQLStats, options *sqlstats.IteratorOptions) TxnStatsIterator {
 	appNames := s.getAppNames(options.SortedAppNames)
 
-	return &TxnStatsIterator{
+	return TxnStatsIterator{
 		baseIterator: baseIterator{
 			sqlStats: s,
 			idx:      -1,
@@ -104,7 +104,7 @@ func NewTxnStatsIterator(s *SQLStats, options *sqlstats.IteratorOptions) *TxnSta
 func (t *TxnStatsIterator) Next() bool {
 	// If we haven't called Next() for the first time or our current child
 	// iterator has finished iterator, then we increment s.idx.
-	if t.curIter == nil || !t.curIter.Next() {
+	if !t.curIter.Initialized() || !t.curIter.Next() {
 		t.idx++
 		if t.idx >= len(t.appNames) {
 			return false