--- a/pkg/sql/sqlstats/ssmemstorage/ss_mem_iterator.go
+++ b/pkg/sql/sqlstats/ssmemstorage/ss_mem_iterator.go
@@ -33,7 +33,7 @@ type StmtStatsIterator struct {
 // NewStmtStatsIterator returns a StmtStatsIterator.
 func NewStmtStatsIterator(
 	container *Container, options *sqlstats.IteratorOptions,
-) *StmtStatsIterator {
+) StmtStatsIterator {
 	var stmtKeys stmtList
 	container.mu.Lock()
 	for k := range container.mu.stmts {
@@ -44,13 +44,18 @@ func NewStmtStatsIterator(
 		sort.Sort(stmtKeys)
 	}
 
-	return &StmtStatsIterator{
+	return StmtStatsIterator{
 		baseIterator: baseIterator{
 			container: container,
 			idx:       -1,
 		},
 		stmtKeys: stmtKeys,
 	}
+}
+
+// Initialized returns true if the iterator has been initialized, false otherwise.
+func (s *StmtStatsIterator) Initialized() bool {
+	return s.container != nil
 }
 
 // Next updates the current value returned by the subsequent Cur() call. Next()
@@ -119,9 +124,7 @@ type TxnStatsIterator struct {
 }
 
 // NewTxnStatsIterator returns a new instance of TxnStatsIterator.
-func NewTxnStatsIterator(
-	container *Container, options *sqlstats.IteratorOptions,
-) *TxnStatsIterator {
+func NewTxnStatsIterator(container *Container, options *sqlstats.IteratorOptions) TxnStatsIterator {
 	var txnKeys txnList
 	container.mu.Lock()
 	for k := range container.mu.txns {
@@ -132,13 +135,18 @@ func NewTxnStatsIterator(
 		sort.Sort(txnKeys)
 	}
 
-	return &TxnStatsIterator{
+	return TxnStatsIterator{
 		baseIterator: baseIterator{
 			container: container,
 			idx:       -1,
 		},
 		txnKeys: txnKeys,
 	}
+}
+
+// Initialized returns true if the iterator has been initialized, false otherwise.
+func (t *TxnStatsIterator) Initialized() bool {
+	return t.container != nil
 }
 
 // Next updates the current value returned by the subsequent Cur() call. Next()