This PR contains two changes that each avoid per-transaction heap allocations in the `sqlstats` package.

### sql/sqlstats: return iterators by value from constructors

This commit switches to returning sqlstats iterations by value from their constructors, instead of by pointer. This avoids causing the iterators to escape and needing to allocate on the heap. Instead, they can be kept on the stack, avoiding heap allocations in methods like `IterateStatementStats`.

```sh
➜ benchdiff --run='BenchmarkKV/./SQL/rows=1$$' --count=25 ./pkg/sql/tests

name                     old time/op    new time/op    delta
KV/Insert/SQL/rows=1-10     173µs ±13%     172µs ±15%    ~     (p=0.939 n=25+25)
KV/Update/SQL/rows=1-10     285µs ±16%     280µs ±11%    ~     (p=0.356 n=25+23)
KV/Delete/SQL/rows=1-10     212µs ±25%     221µs ±19%    ~     (p=0.345 n=25+25)
KV/Scan/SQL/rows=1-10       127µs ±11%     127µs ±11%    ~     (p=0.878 n=25+24)

name                     old alloc/op   new alloc/op   delta
KV/Scan/SQL/rows=1-10      32.2kB ± 1%    32.0kB ± 1%  -0.36%  (p=0.006 n=24+24)
KV/Insert/SQL/rows=1-10    59.0kB ± 1%    58.9kB ± 1%    ~     (p=0.140 n=25+25)
KV/Update/SQL/rows=1-10    70.6kB ± 1%    70.4kB ± 1%    ~     (p=0.050 n=24+24)
KV/Delete/SQL/rows=1-10    84.9kB ± 1%    84.9kB ± 1%    ~     (p=0.859 n=25+25)

name                     old allocs/op  new allocs/op  delta
KV/Scan/SQL/rows=1-10         368 ± 1%       366 ± 1%  -0.62%  (p=0.001 n=24+24)
KV/Update/SQL/rows=1-10       726 ± 1%       723 ± 2%  -0.45%  (p=0.013 n=23+24)
KV/Insert/SQL/rows=1-10       492 ± 2%       490 ± 1%  -0.41%  (p=0.044 n=25+25)
KV/Delete/SQL/rows=1-10       613 ± 3%       613 ± 3%    ~     (p=0.179 n=25+25)
```

### sql/sqlstats: pass `sqlstats.IteratorOptions` by value, not pointer

This commit switches from passing the `sqlstats.IteratorOptions` around by pointer to passing it by value. The struct is 2 bytes large (8 when padded), so there's little benefit to passing it around by pointer. Meanwhile, passing the object by pointer through interface methods prevents escape analysis from keeping it on the stack, forcing a heap allocation.

```
➜ benchdiff --run='BenchmarkKV/./SQL/rows=1$$' --count=25 ./pkg/sql/tests

name                     old time/op    new time/op    delta
KV/Insert/SQL/rows=1-10     169µs ±13%     168µs ±17%    ~     (p=0.603 n=25+25)
KV/Update/SQL/rows=1-10     282µs ± 9%     282µs ± 8%    ~     (p=0.788 n=25+25)
KV/Delete/SQL/rows=1-10     227µs ±26%     206µs ±27%    ~     (p=0.074 n=25+25)
KV/Scan/SQL/rows=1-10       126µs ±12%     127µs ±16%    ~     (p=0.908 n=25+25)

name                     old alloc/op   new alloc/op   delta
KV/Delete/SQL/rows=1-10    84.9kB ± 1%    84.5kB ± 1%  -0.50%  (p=0.008 n=25+21)
KV/Insert/SQL/rows=1-10    58.9kB ± 1%    58.7kB ± 1%  -0.23%  (p=0.020 n=25+23)
KV/Update/SQL/rows=1-10    70.5kB ± 1%    70.5kB ± 1%    ~     (p=0.894 n=24+24)
KV/Scan/SQL/rows=1-10      32.0kB ± 1%    32.0kB ± 1%    ~     (p=0.631 n=25+24)

name                     old allocs/op  new allocs/op  delta
KV/Delete/SQL/rows=1-10       613 ± 3%       604 ± 1%  -1.52%  (p=0.000 n=25+21)
KV/Insert/SQL/rows=1-10       489 ± 1%       486 ± 1%  -0.69%  (p=0.001 n=25+23)
KV/Scan/SQL/rows=1-10         365 ± 1%       363 ± 1%  -0.42%  (p=0.024 n=25+25)
KV/Update/SQL/rows=1-10       724 ± 2%       722 ± 2%    ~     (p=0.302 n=25+24)
```

Epic: None
Release note: None